api/fastapi_sms_app/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
python-dotenv>=1.0
twilio>=9.0
